refactor: Renamed client module from ollamaClient to ollama
Updated import statements in commands.js to use new module name
